Moves old files. Updates TestPopupMessage
